% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetperfectum_vraag_imperfectum_sing.R
\name{hetperfectum_vraag_imperfectum_sing}
\alias{hetperfectum_vraag_imperfectum_sing}
\title{Spel begin}
\usage{
hetperfectum_vraag_imperfectum_sing()
}
\value{

}
\description{
Spel begin
}
