% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetperfectum.R
\docType{data}
\name{hetperfectum}
\alias{hetperfectum}
\title{Deze dataset heeft de infinitief een het perfectum voor veel verbs}
\format{
Een tibble (tidyverse data frame)
}
\usage{
hetperfectum
}
\description{
Deze dataset heeft de infinitief een het perfectum voor veel verbs
}
\keyword{datasets}
